캐시되지 않은 파일의 캐시를 삭제하려고 시도하던 버그 수정

+++ src/lib/modules/cache.ts
@@ -42,7 +42,9 @@ export const storeFileCache = async (fileId: number, fileBuffer: ArrayBuffer) =>
 };
 
 export const deleteFileCache = async (fileId: number) => {
-  await deleteFile(`/cache/${fileId}`);
+  if (!fileCacheIndex.has(fileId)) return;
+
   fileCacheIndex.delete(fileId);
+  await deleteFile(`/cache/${fileId}`);
   await deleteFileCacheIndex(fileId);
 };
